# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MPI API Documentação",
        "description": "Documentação da API do sistema MPI",
        "contact": {
            "email": "luizleao@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tipo-ator": {
            "get": {
                "tags": [
                    "TiposAtor"
                ],
                "summary": "Listar todos os tipos de ator",
                "operationId": "69c2f7a62c8df6d683f04f8eb21187d3",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de ator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoAtor"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TiposAtor"
                ],
                "summary": "Criar um novo tipo de ator",
                "operationId": "aec4e802ea378a3cc1f9192dfb4ecbb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Tipo de Ator do Conflito"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de Ator criado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/tipo-inquerito-policial": {
            "get": {
                "tags": [
                    "TiposInqueritoPolicial"
                ],
                "summary": "Listar todos os tipos de Inquerito Policial",
                "operationId": "1567920eeb77f861d9d475fb01841943",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de Inquerito Policial",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoInqueritoPolicial"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TiposInqueritoPolicial"
                ],
                "summary": "Criar um novo tipo de Inquerito Policial",
                "operationId": "d0dc250ef860c3dca9a116ba22453224",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Tipo de Inquerito Policial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de Inquerito Policial criado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/tipo-ator/{id}": {
            "get": {
                "tags": [
                    "TiposAtor"
                ],
                "summary": "Obter um tipo de ator específico",
                "operationId": "21e88a7a38f1d56713df00b423bca08b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do tipo de ator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoAtor"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TiposAtor"
                ],
                "summary": "Atualizar um tipo de ator específico",
                "operationId": "77e8679c1e49d163a12cabe5ec4bfbc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Tipo de Ator do Conflito"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de Ator atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TiposAtor"
                ],
                "summary": "Excluir um tipo de ator específico",
                "operationId": "30a4f6415546cddedcdab3ded492f1e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tipo de Ator excluído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/tipo-inquerito-policial/{id}": {
            "get": {
                "tags": [
                    "TiposInqueritoPolicial"
                ],
                "summary": "Obter um tipo de Inquerito Policial específico",
                "operationId": "d37bb6c6cbfc2754b8f77a3b1f7befa0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do tipo de Inquerito Policial",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoInqueritoPolicial"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "TiposInqueritoPolicial"
                ],
                "summary": "Atualizar um tipo de Inquerito Policial específico",
                "operationId": "e9b8d703b8a89a847e3d8ded38480b12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Tipo de Inquerito Policial"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de Inquerito Policial atualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TiposInqueritoPolicial"
                ],
                "summary": "Excluir um tipo de Inquerito Policial específico",
                "operationId": "1d7c982a628a48316c210fac54469172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tipo de Inquerito Policial excluído"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TipoAtor": {
                "properties": {
                    "idTipoAtor": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Latifundiário"
                    }
                },
                "type": "object"
            },
...
            "TipoInqueritoPolicial": {
                "properties": {
                    "idTipoInqueritoPolicial": {
                        "type": "integer",
                        "example": "3"
                    },
                    "nome": {
                        "type": "string",
                        "example": "Processos em tramitação na justica federal"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}